+++ src/aa_device_save_data.cu
@@ -1,3 +1,6 @@
+#include <iostream>
+#include <stdio.h>
+
 namespace astroaccelerate {
   /** \brief Copy data and set up the GPU constants/variables. */
   void save_data(float *device_pointer, float *host_pointer, size_t size) {
@@ -5,7 +8,7 @@ namespace astroaccelerate {
   }
 
   /** \brief Copy data and set up the GPU constants/variables. */
-  void save_data_offset(float *device_pointer, int device_offset, float *host_pointer, int host_offset, size_t size) {
-    cudaMemcpy(host_pointer + host_offset, device_pointer + device_offset, size, cudaMemcpyDeviceToHost);
+  void save_data_offset(float *device_pointer, size_t device_offset, float *host_pointer, size_t host_offset, size_t size) {
+	cudaMemcpy(host_pointer + host_offset, device_pointer + device_offset, size, cudaMemcpyDeviceToHost);
   }
 } //namespace astroaccelerate
